🚚 Rename: utils, constants 폴더 생성, 폴더 구조 변경
